chore(web/applications): remove now unnecessary MSALPlugin class
